requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
babel==2.16.0
bcrypt==4.2.0
click==8.1.7
fastapi==0.115.2
fastapi-admin==1.0.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
pandas==2.2.3
pendulum==3.0.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==1.7.1
pypika-tortoise==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sqladmin==0.20.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.40.0
time-machine==2.16.0
tortoise-orm==0.21.7
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
whitenoise==6.8.0
WTForms==3.1.2
XlsxWriter==3.2.0
